requirements.txt
faiss-cpu
langchain
langchain-community
langchain-openai
langgraph
langsmith
pandas
pandasql
python-dotenv
uuid
